Tema7/requirements.txt
Django>=5.2
djangorestframework
drf-yasg
django-cors-headers
psycopg2-binary
gunicorn
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.6
django-ckeditor==6.7.3
django-js-asset==3.1.2
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
sqlparse==0.5.3
